example/pytest_factory_async_repro/requirements.txt
django>=5.0
factory-boy>=3.3
pytest>=8.0
pytest-asyncio>=0.24
pytest-django>=4.11
pytest-django-asyncio>=0.1.5
pytest-factoryboy>=2.7
